EG-49: IzPack Installer. Added input panels requesting EC2 credentials, EC2 keypair and EG S3 drop bucket.
Currently there is no validation of the inputs (except that they are filled in) but the eg.properties file is properly generated based on the user input.

The installer should allow creation of EC2 keypairs and retreive the list of available keypairs in the AWS account. The same should be done for S3 buckets.

git-svn-id: svn+ssh://svn.forge.objectweb.org/svnroot/elastic-grid/trunk@261 e25e3cd6-ff2c-412d-a5cc-22119c6ca403